move util folder (patches update)
