+++ src/components/views/rooms/TchapExternalRoomHeader.tsx
@@ -4,7 +4,7 @@ import React, { useState } from "react";
 import { _t } from "matrix-react-sdk/src/languageHandler";
 
 import { TchapRoomAccessRule, TchapRoomAccessRulesEventId } from "../../../tchap/@types/tchap";
-import TchapRoomUtils from "../../../util/TchapRoomUtils";
+import TchapRoomUtils from "../../../tchap/util/TchapRoomUtils";
 
 export interface IProps {
     room?: Room;
