+++ src/components/views/settings/TchapJoinRuleSettings.tsx
@@ -41,7 +41,7 @@ import { doesRoomVersionSupport, PreferredRoomVersions } from "matrix-react-sdk/
 import LabelledToggleSwitch from "matrix-react-sdk/src/components/views/elements/LabelledToggleSwitch";
 import QuestionDialog from "matrix-react-sdk/src/components/views/dialogs/QuestionDialog";
 
-import TchapUIFeature from "../../../util/TchapUIFeature";
+import TchapUIFeature from "../../../tchap/util/TchapUIFeature";
 import { TchapRoomAccessRule, TchapIAccessRuleEventContent, TchapRoomAccessRulesEventId } from "../../../tchap/@types/tchap";
 
 interface IProps {
