+++ src/tchap/lib/ExpiredAccountHandler.ts
@@ -7,7 +7,7 @@ import Modal from "matrix-react-sdk/src/Modal";
 import PlatformPeg from "matrix-react-sdk/src/PlatformPeg";
 
 import ExpiredAccountDialog from "../../components/views/dialogs/ExpiredAccountDialog";
-import TchapUtils from "../../util/TchapUtils";
+import TchapUtils from "../util/TchapUtils";
 
 /*
  * Listens for HttpApiEvent.ORG_MATRIX_EXPIRED_ACCOUNT events and opens the panel ExpiredAccountDialog.
