+++ src/tchap/util/TchapStrongPassword.tsx
